feat: useErrorStatus 훅 구현
